requirements.txt
Flask
Pillow
pytesseract
gTTS
langdetect
indic-transliteration
pyttsx3